+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.0] - 12-29-23
+
+- Added toggle object support to Access Keypad Controller
+- Added JSON support to Remote Whitelist
+- Added UserList API property to Static Whitelist
+- Updated Whitelist Grant to infer target Access Control object if not set
+- Updated implementation of Forwarding Handler to extend from access handler base class
+- Added prefab for forwarding access handler
+
 ## [1.0.2] - 09-29-23
 
 - Removed dependency on UdonSharp package to prepare for SDK 3.4.0